platform: nordic_nrf: Place a few .rodata sub-sections before veneers
Modify the nRF specific linker script for the secure image so that
two sections with read-only data (containing arrays service_db[] and
memory_data_list[]) are placed before veneers (and consume part of
the gap resulting from the veneers section alignement).
Without this patch, the RegressionIPC configuration fails to build
because of the FLASH overflow (.TFM_DATA cannot fit).

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>