platform: nordic_nrf: Improve definition of NRFX_ASSERT()
When the NDEBUG flag is used in compilation, assert() becomes empty.
This may lead to warnings about unused variables or functions if those
are used only in the asserted expression.
Instead of basing NRFX_ASSERT() directly on assert() for NDEBUG case,
use an if statement that will swallow the assertion condition without
evaluating it.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>